% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/AR.root.R
\name{AR.root}
\alias{AR.root}
\title{Computing roots of the characteristic AR polynomial}
\usage{
AR.root(Arima)
}
\arguments{
\item{Arima}{Fitted ARIMA model from \code{\link{arima}}}
}
\value{
returns a three column data frame with the real,
  the imaginary part and the radius of the roots.
  The number of rows corresponds to the coefficients.
}
\description{
Compute roots of the characteristic AR polynomial.
}

