% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arimaMisc.R
\docType{package}
\name{arimaMisc}
\alias{arimaMisc}
\alias{arimaMisc-package}
\title{arimaMisc: Time Series Analysis Miscellaneous Tools}
\description{
This is a package collecting some useful function for Time Series Analysis, 
including statistical tests for time series data and several functions 
improving ARIMA modeling experience.
}

