% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arimaPred.plot.R
\name{arimaPred.plot}
\alias{arimaPred.plot}
\title{Plot the prediction and confidence interval of arima model}
\usage{
arimaPred.plot(series, Arima, b, h, level = 0.95, new.xreg = NULL)
}
\arguments{
\item{series}{time series data}

\item{Arima}{Fitted ARIMA model from \code{\link{arima}}}

\item{b}{number of periods for original data plotting}

\item{h}{number of periods for forecasting}

\item{level}{confidence level for prediction intervals}

\item{new.xreg}{optional, a vector or matrix of external regressors}
}
\description{
Compute the prediction and standard error of the prediction and
plot the prediction and confidence interval of arima model
}

