% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/armaSigOrder.R
\name{armaSigOrder}
\alias{armaSigOrder}
\title{Generating fixed parameter for an ARIMA full model}
\usage{
armaSigOrder(Arima)
}
\arguments{
\item{Arima}{Fitted ARIMA model from \code{\link{arima}}}
}
\value{
fixed parameter for an arima function's fixed argument
}
\description{
Performe Z test for arma model's coefficients and Return an vector of
fixed parameter for \code{\link{arima}} function's fixed argument.
}

