% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/auto.cov.R
\name{auto.cov}
\alias{auto.cov}
\title{Compute autocovariance function of ARMA process}
\usage{
auto.cov(ar = 0, ma = 0, sigma2 = 1, lag)
}
\arguments{
\item{ar}{AR parameters}

\item{ma}{MA parameters}

\item{sigma2}{sigma2}

\item{lag}{lag}
}
\value{
parameters of autocovariance function
}
\description{
Compute autocovariance function of ARMA process.
}

