% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Box.plot.R
\name{Box.plot}
\alias{Box.plot}
\title{Plot the Ljung-Box Test's p-value vs to lag}
\usage{
Box.plot(Arima, h = 48)
}
\arguments{
\item{Arima}{Fitted ARIMA model from \code{\link{arima}}}

\item{h}{maximum lag}
}
\description{
Compute and Plot the Ljung-Box Test's p-value vs lag.
}

