% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Box.Test.R
\name{Box.Test}
\alias{Box.Test}
\title{Corrected Ljung-Box Test for ARIMA model's residuals}
\usage{
Box.Test(Arima, lag = 12)
}
\arguments{
\item{Arima}{Fitted ARIMA model from \code{\link{arima}}}

\item{lag}{the statistic based on \emph{lag} autocorrelation coefficients}
}
\value{
p-value of Ljung-Box Test
}
\description{
Compute Ljung Box test statistic and return p-value for examining an ARIMA
model's residuals, with correct degree of freedom. This is an efficient
version of \code{\link{Box.test}} for testing ARIMA model's residuals.
}

