% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EWMA.Cov.R
\name{EWMA.Cov}
\alias{EWMA.Cov}
\title{Compute the exponentially weighted moving average covariance matrix}
\usage{
EWMA.Cov(rtn, theta = 0.94)
}
\arguments{
\item{rtn}{return series}

\item{theta}{exponetial decay factor between 0 and 1}
}
\value{
exponentially weighted moving average covariance matrix
}
\description{
Compute the exponentially weighted moving average covariance matrix.
}

