% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurtosis.test.R
\name{kurtosis.test}
\alias{kurtosis.test}
\title{Testing normality based on fourth central moment}
\usage{
kurtosis.test(series, sign = 0.05)
}
\arguments{
\item{series}{time series data}

\item{sign}{significant level}
}
\value{
p-value
}
\description{
Test normality based on asymptotic properties of fourth central moment
}

