% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MA.root.R
\name{MA.root}
\alias{MA.root}
\title{Computing roots of the characteristic MA polynomial}
\usage{
MA.root(Arima)
}
\arguments{
\item{Arima}{Fitted ARIMA model from \code{\link{arima}}}
}
\value{
returns a three column data frame with the real, 
  the imaginary part and the radius of the roots.
  The number of rows corresponds to the coefficients.
}
\description{
Compute roots of the characteristic MA polynomial
}

