% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/memeory.R
\name{memory}
\alias{memory}
\title{Compute memory function of ARMA process}
\usage{
memory(ar = 0, ma = 0, lag)
}
\arguments{
\item{ar}{AR parameters}

\item{ma}{MA parameters}

\item{lag}{lag}
}
\value{
parameters of memory function
}
\description{
Compute memory function of ARMA process.
}

