% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Ridge.IC.R
\name{Ridge.IC}
\alias{Ridge.IC}
\title{Select ridge regression parameter based on AIC or BIC}
\usage{
Ridge.IC(formula, data, lambda, IC = "AIC")
}
\arguments{
\item{formula}{a formula expression as for regression models, of the form
response ~ predictors. See the documentation of formula for other details.}

\item{data}{an optional data frame in which to interpret the variables
occuring in formula}

\item{lambda}{A ridge regression parameter, May be a vector}

\item{IC}{character, indicating which information criterion to use,
"AIC" or "BIC"}
}
\value{
optimal ridge regression parameter
}
\description{
Select ridge regression parameter based on AIC or BIC.
}

