% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rolling.forecast.ase.R
\name{rolling.forecast.ase}
\alias{rolling.forecast.ase}
\title{Computing the out-of-sample forecasting squared error of rolling predictions}
\usage{
rolling.forecast.ase(series, Arima, range = 10, start = 50, plot = FALSE)
}
\arguments{
\item{series}{time series data}

\item{Arima}{Fitted ARIMA model from \code{\link{arima}}}

\item{range}{the forecasting range}

\item{start}{starting point for out-of-sample forecast}

\item{plot}{logical flag: if TRUE, then plot the observations v.s. prediction}
}
\value{
mean square error of the prediction
}
\description{
Calculating the average squared error of one step rolling predictions.
}

