% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skewness.test.R
\name{skewness.test}
\alias{skewness.test}
\title{Testing normality based on third central moment}
\usage{
skewness.test(series, sign = 0.05)
}
\arguments{
\item{series}{time series data}

\item{sign}{significant level}
}
\value{
p-value
}
\description{
Test normality based on asymptotic properties of third central moment.
}

